requirements.txt
annotated-types==0.7.0
asn1crypto==1.5.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
id==1.4.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.1
pg8000==1.31.2
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
python-dateutil==2.9.0.post0
requests==2.32.3
scramp==1.4.5
six==1.16.0
SQLAlchemy==2.0.36
testing.common.database==2.0.3
testing.postgresql==1.3.0
tomli==2.0.2
typing_extensions==4.12.2
urllib3==2.2.3
waitress==3.0.0
Werkzeug==3.0.6
zipp==3.20.2